Finalizado o tópico: Isolamento de imports
close #28
